Merged PR 1170: add an optional "nfft" argument to FFTNode that allows you to change the number
Add an optional "nfft" argument to FFTNode that allows you to change the number of frequencies we look for in the input (and therefore also changes the size of the output to nfft/2).  The numpy fft.rfft function has this option.

With this pull request I can now match the python_speech_features featurizer because they are using the default value of nfft=512.  To test this all I had to do was change my TestFFT.ipynb to add the new nfft parameter and the output now matches perfectly even when the input size is 160, because nfft is 256 which is a power of 2.

Related work items: #1681