Merged PR 1181: Fix importing for densenets
Fix importing for densenets

This has two fixes:
- In the ONNX reshape node converter, once the input and output shapes
are set, remove any input tensors so that "Skip" nodes that govern the
shape aren't in the input set for the "Reshape" node any longer. This
enables the "Reshape" node to be included in the ordered list of
imported nodes
- When constructing an ell::model::Map with the converted ell nodes, a
Submodel is constructed holding the ell model. In the Submodel
constructor we call Submodel::VerifyInputs to verify that all inputs can
be seen, but this was examining each possible route from the end of the
model to the beginning, which grows nearly exponentially with the
number of dense layers in a densenet, so it was incredibly slow.
To fix this, maintain a set of visited nodes and only visit each
node once.