Merged PR 1234: Fix bugs in IRHeaderWriter to do with callbacks and duplicate method args
Call backs can create members with duplicate names, which won't compile.  This fixes that by prefixing the member names with the method names.  Also fixed a related problem in IRHeaderWriter where it contains some brittle code that depended on the name of output arguments to generate the correct wrapper.  But the name of arguments can now come from metadata in the model from the importer, so this doesn't work for some of our CNTK models (wrapper won't compile).  This fix means plumbing information from MapCompiler so IRModuleEmitter knows which parameters are inputs and which are outputs.  This information is then passed to the IRHeaderWriter and everything is good and no longer dependent on user defined metadata.  And I added a new unit test to cover all these bugs in the IRHeaderWriter.