Merged PR 1260: Fix issues with UnaryOperationNode<int> and the "square" operation
This PR kind-of fixes a problem with UnaryOperationNode<int> and the "square" operation always returning zero.

The fix isn't much of a fix, but rather a workaround. The real problem had to do with the expression `x * x` not generating code when `x` was a value library `Scalar` of type int.

Added a work item to fix the underlying problem, though. #2185

Related work items: #2155