Merged PR 1280: Couple of fixes for value::LLVMContext::CopyDataImpl
Couple of fixes for value::LLVMContext::CopyDataImpl

This change makes a couple of fixes for `CopyDataImpl` in
`value::LLVMContext`:
* When the destination is constant, it will automatically promote the
destination and store the result. This is needed in something like the
following
```cpp
Scalar x = 2;
x = x * y; // assuming y is some previously defined non-constant value
```
* Correctly calculate the size of the type when calculating the memory
size to copy. We were using the size of the pointer to the type instead
of the type itself (equivalent to doing `sizeof(int*)` instead of
`sizeof(int)`)