Merged PR 1334: Improvements to enable LLVM's vectorizer during optimization
It turns out LLVM wasn't able to auto-vectorize generated code, due a missing analysis pass in the optimization pass pipeline. This PR adds it, plus a number of other passes that `opt` enables.

This PR also adds `llvm.loop` metadata to the conditional branch instruction on "for" and "while" loops. Later we can add more properties to that metadata to control LLVM's vectorization or unrolling or whatever.