Merged PR 1348: Fixed issue with python bindings on a Mac
The python interpreter installed via `conda` statically links the python libraries, resulting in a segfault when importing the ELL python bindings (be we dynamically link to the python libraries). This PR fixes the issue (for me, at least).

The CMake code that builds the python bindings should really have a more robust method of detecting when to dynamically link to the python libs and when not to. Right now it omits linking if the CMake `APPLE` config variable is set.