Merged PR 1357: Add support for reference semantics to Value library
This change adds two important member functions to `ell::value::Value`, `Reference() -> Value` and `Dereference() -> Value`. These are equivalent to taking the address of the value and dereferencing the address specified at the value, respectively.

A helper class, `ell::value::Ref<T>` has also been added to provide a cleaner interface. `Ref<T>` can be constructed such that it automatically stores the reference to the passed in value provided to the constructor. It also provides an implicit conversion operator for `T`, which when invoked will automatically dereference the stored value.