Merged PR 1359: Use a custom type alias instead of intptr_t or uintptr_t
Use a custom type alias instead of `intptr_t` or `uintptr_t`

macOS seems to alias `intptr_t` to `long`, which is different from what they
alias for `int32_t` and `int64_t`, which are `int` and `long long`,
respectively. We will now use a custom type to alias to either `int32_t`
or `int64_t`, depending on the value of `sizeof(void*)`. If neither
`int32_t` nor `int64_t` match the width of `void*`, then we will raise a
`static_assert` and fail to compile.