Merged PR 1402: Fix a codegen error with local variables
Local variables ("`alloca`s") in LLVM functions are supposed to all appear in the first basic block of a function (the _entry block_), and before any other code. The emitters library only got this half right: all `alloca` instructions were placed in the entry block, but they could be interleaved with other code in that block.

This PR fixes the issue: now the entry block contains only `alloca`s. Emitted code begins in the second basic block of the function. There's a test that catches the error (and now passes).

There's an unrelated change that removes a confusing parameter to `IRFunctionEmitter::Variable()`.

Related work items: #2389