Merged PR 1451: Directly store zero for scalars
This PR contains a minor change in how values are initialized to zero.

Previously, values were zeroed out by calling `memset`. In this version , if the number of elements to initialize is 1, the value is stored directly. The motivation for this change is just to improve readability of the output IR. It helps a little bit.

Example of old output:
```
  %0 = alloca i32
  %1 = bitcast i32* %0 to i8*
  call void @llvm.memset.p0i8.i64(i8* %1, i8 0, i64 4, i1 false)
```

New output:
```
  %0 = alloca i32
  store i32 0, i32* %0
```