Dockerfiles for MS Container Registry
Introducing multi-staged Dockerfiles for publishing ELL-ready images to MS Container Registry.

These are currently building in an Azure Pipelines task, and automatically pushing to ellcr.azurecr.io container registry. This is linked to MCR (Microsoft Container Registry) so, when we're ready to flip the switch, these images will be available to the public so they can use ELL without installing all the tools. These can be run locally or as an Azure Batch instance -- instructions for both solutions will be provided.